move int conversion types from util_math to types adapt vector to new functions
