+++ include/dlaf/common/vector.h
@@ -11,6 +11,7 @@
 #pragma once
 
 #include <vector>
+#include "dlaf/types.h"
 
 namespace dlaf {
 namespace common {
@@ -18,14 +19,14 @@ namespace common {
 /// std::vector with overloads for signed indexes
 template <typename T>
 struct vector : public std::vector<T> {
-  vector(int size) : std::vector<T>(static_cast<std::size_t>(size)) {}
+  vector(int size) : std::vector<T>(to_sizet(size)) {}
 
   T& operator[](int index) {
-    return std::vector<T>::operator[](static_cast<std::size_t>(index));
+    return std::vector<T>::operator[](to_sizet(index));
   }
 
   const T& operator[](int index) const {
-    return std::vector<T>::operator[](static_cast<std::size_t>(index));
+    return std::vector<T>::operator[](to_sizet(index));
   }
 };
